move resolvconf updater to its own file, ensure newlines on end of all lines
